+++ test/Fixture/Methods/NoReturnByReferenceRule/MethodInInterfaceReturningByReference.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ergebnis\PHPStan\Rules\Test\Fixture\Methods\NoReturnByReferenceRule;
+
+interface MethodInInterfaceReturningByReference
+{
+    public function &foo($bar);
+}